[refactor](be) Move column-filling logic from GenericReader to TableFormatReader
### What problem does this PR solve?

Problem Summary: Column-filling state (partition values, missing defaults,
synthesized column handlers) and methods (on_fill_partition_columns,
on_fill_missing_columns, fill_remaining_columns, etc.) were in GenericReader
but only used by table-level readers. This polluted the base class interface.

Move all column-filling data members, methods, and the default
on_before_init_reader implementation from GenericReader to TableFormatReader:

- set_fill_column_data, missing_cols, on_fill_partition_columns,
  on_fill_missing_columns, fill_synthesized_columns, fill_remaining_columns,
  register_synthesized_column_handler, _extract_partition_values
- Data: _fill_partition_values, _fill_missing_defaults,
  _fill_col_name_to_block_idx, _fill_missing_cols, _synthesized_col_handlers

GenericReader::on_before_init_reader is now a minimal no-op (just sets
_column_descs). RowGroupReader pointer changed from GenericReader* to
TableFormatReader*.

### Release note

None

### Check List (For Author)

- Test: No need to test - pure code movement, no behavior change
- Behavior changed: No
- Does this need documentation: No
